backend/requirements.txt
Django==5.2.6
djangorestframework~=3.15
djangorestframework-simplejwt~=5.3
django-cors-headers~=4.4
Pillow~=10.4
cryptography~=43.0
bcrypt~=4.2
argon2-cffi~=23.1
gunicorn~=22.0
psycopg2-binary~=2.9
whitenoise~=6.6
